requirements.txt
# Backend dependencies
fastapi
uvicorn
httpx
pydantic
python-dotenv
requests
python-multipart
# Front end dependencies
streamlit
requests